Listener for receiving data implemented. Service now works with connector and updates data on devices themselves.

+++ src/main/java/com/example/zigbeepolivalka/domain/MoistureMode.java
@@ -12,7 +12,7 @@ public class MoistureMode extends AbstractMode {
   }
 
   @Override
-  public int getModeId() {
+  public byte getModeId() {
     return 1;
   }
